Non spamming logs feature
For not getting repeated logs, a Json folder is created, with each object's name as the json files name, which stores the last log sent for each image, and then compares. If the last sent log was the same as the one that is about to be sent now, it is not sent. This is specially useful when just nothing happens, not to have a log full of 'Nothing to change' messages